+++ osafw-app/App_Code/fw/FwDynamicController.cs
@@ -73,7 +73,9 @@ public partial class FwDynamicController : FwController
             setViewList();
 
         // set standard output parse strings
+#pragma warning disable CS0618 // Preserve dispatch to existing project overrides of setPS().
         var ps = this.setPS();
+#pragma warning restore CS0618
 
         // userlists support if necessary
         if (this.is_userlists)